Add auth to product mutations

+++ app/product/mutations/deleteProduct.ts
@@ -1,4 +1,4 @@
-import { resolver } from 'blitz'
+import { AuthorizationError, Ctx, resolver } from 'blitz'
 import db from 'db'
 import { z } from 'zod'
 
@@ -9,8 +9,8 @@ const DeleteProduct = z.object({
 const deleteProduct = resolver.pipe(
   resolver.zod(DeleteProduct),
   resolver.authorize(),
-  async ({ id }) => {
-    // TODO: in multi-tenant app, you must add validation to ensure correct tenant
+  async ({ id }, { session }: Ctx) => {
+    if (!session.userId) throw new AuthorizationError()
     const product = await db.product.deleteMany({ where: { id } })
 
     return product
